Move into redb_database module
